+++ framework/platform-dx/rasterizerstate.cpp
@@ -72,7 +72,7 @@ namespace xna {
 	}
 
 	void RasterizerState::DepthBias(float value) {
-		impl->dxDescription.DepthBias = value;
+		impl->dxDescription.DepthBias = static_cast<INT>(value);
 	}
 
 	float RasterizerState::SlopeScaleDepthBias() const {
